// src/app/modul/admin/teacher/models/teachers.model.ts
export interface TeacherResponse {
  id: number;
  name: string;
  description: string;
  addres: string;
  dataOfBirth: string;
  dataOfRegister: string;
  phone: string;
  email: string;
  telegramUserName: string;
  specialization: string;
}
export interface TeacherResquest {
  name: string;
  description: string;
  addres: string;
  dataOfBirth: string;
  dataOfRegister: string;
  phone: string;
  email: string;
  telegramUserName: string;
  specialization: string;
}